requirements.txt
flask==3.0.0
playwright==1.40.0
openai==1.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.26.2
Pillow==10.1.0
pypdf2==3.0.1
pdfplumber==0.10.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
openpyxl==3.1.2
pytesseract==0.3.10
opencv-python==4.8.1.78
scipy==1.11.4
scikit-learn==1.3.2
networkx==3.2.1
geopandas==0.14.1
gunicorn==21.2.0
